functions: create a decompress_cat() function
And move the kver_zimage function into functions.bats.  The only
reason it still exists at this point is because some of the tests use
it and I don't care to rewrite the tests.  Users won't be broken by
this because if they were, we'd have already broken them a few commits
ago when we added another argument to the function.