Use /sysroot as the root mountpoint
This way both busybox and systemd based initramfs will use the same
mountpoint.

The change is mostly cosmetical, but it will affect third-party hooks
that interact with /new_root.
To avoid issues (or to at least try to avoid them), /new_root will be
kept as a symlink to /sysroot.