mkinitcpio: do not try to decompress nonexistent firmware
If the firmware directories do not exist, the find command gets passed an
empty array making it search in the current directory instead. This
results in uncompressing all `.zstd` files in the current directory tree.

Make sure `decompress_firmware` does not run if there are no firmware
directories and make sure `find` is passed a directory to seach in.

Fixes: 4dd9fdedbc2b8f5a5cce0f2882816a2f9f7c4ce6 ("mkinitcpio: use MODULES_DECOMPRESS to also control firmware file decompression")