Revert "install/keyboard: add more HID modules"
This reverts commit 27039f7a3a2aeebd1846f47fac728d70dddd13db.

The availability of `i2c-hid*` modules in early userspace breaks
touchpad on Framework 16 laptops.

This is without doubt either a firmware issue or a kernel bug, but since
no one actually asked for the change to include these modules to be made,
it should be fine to simply revert it.

See https://community.frame.work/t/touchpad-not-working-since-update-archlinux/50304

Closes https://gitlab.archlinux.org/archlinux/mkinitcpio/mkinitcpio/-/issues/265