install/block: include more low level drivers to fix pci/usb controllers
On certain RISC-V boards (and I think generally applied to more embedded devices), there are more low-level drivers that need to be present in the initrd for pci/usb buses to be initialized successfully.

Fixes boot on JH7110-based boards (VisionFive v2 and Milk-V Mars), and
maybe others that shows these errors:

```
[   13.066149] /soc/pcie@2b000000: Failed to get clk index: 1 ret: -517
[   13.066321] /soc/pcie@2c000000: Failed to get clk index: 1 ret: -517
[   13.066386] platform 17020000.pinctrl: deferred probe pending: starfive-jh7110-aon-pinctrl: could not get reset
[   13.066401] platform 2b000000.pcie: deferred probe pending: pcie-starfive: failed to get pcie clocks
[   13.066412] platform 2c000000.pcie: deferred probe pending: pcie-starfive: failed to get pcie clocks
```

These three folders are 644K, 120K, 60K in size, shouldn't be too bad
for the fallback initrd to be enabled generally.