install/systemd: add libacl as dependency
Since systemd v259, libacl will now be opened via dlopen(), so it is no
longer automatically detected as a dependency. This is still used by
systemd-tmpfiles if ACLs are specified in one of its configs, which is
the case for the default systemd tmpfiles. Missing libacl.so will cause
warnings to be written to the log.